Add rehashing based on load factor.
Growth happens on insert, shrinking happens on removal.

The test cases for these assume things about the implementation that
they probably shouldn't. The actual property we'd like to prove is that
no sequence of insertions and removals will cause the hash table to lose
data, so we should really be using fuzz testing.